[REFACT] Service에 있는 메서드를 퍼사드 패턴으로 분리, 매핑 로직 분리, 인덱스에 대한 변수 추가
